Update Notes to be first class collection of Note objects
- Chords now just becomes a specialization of Notes with 2 or more notes
